rename dirs since it was confusing
