+++ modules/cf-core/src/methods/state-channel/get-all-addresses/controller.ts
@@ -1,0 +1,20 @@
+import { jsonRpcMethod } from "rpc-server";
+
+import { RequestHandler } from "../../../request-handler";
+import { Node } from "../../../types";
+import { NodeController } from "../../controller";
+
+export default class GetAllChannelAddressesController extends NodeController {
+  @jsonRpcMethod(Node.RpcMethodName.GET_CHANNEL_ADDRESSES)
+  public executeMethod = super.executeMethod;
+
+  protected async executeMethodImplementation(
+    requestHandler: RequestHandler
+  ): Promise<Node.GetChannelAddressesResult> {
+    return {
+      multisigAddresses: [
+        ...(await requestHandler.store.getStateChannelsMap()).keys()
+      ]
+    };
+  }
+}
